Extract filter engine and deduplicate nested sort comparator
Move the recipe filter/search engine from crafting_gui.cpp into
crafting_gui_helpers so both the curses UI and the future ImGui
implementation can share it.

- filter_recipes() parses comma-separated prefix queries and delegates
  to recipe_subset::reduce()
- query_is_yes() moves as file-local static in the helpers .cpp
- SearchPrefix struct, prefixes vector, and filter_help_start
  translation move as extern const declarations
- filter_crafting_recipes() (the popup dialog) stays in crafting_gui.cpp
  since it has UI coupling

Also fix the duplicated sort comparator in nested recipe expansion:

- Rename recursively_expance_recipes -> recursively_expand_recipes
- Replace inline sort lambda with recipe_sort_compare()

Guard recipe_subset::search() cancelability check with test_mode so
unit tests can exercise reduce()/search() without throwing.

New test cases cover name, exclude, component, primary skill, skill,
difficulty range, comma chaining, memorized yes/no, book source, and
two characterization tests for pre-existing quirks (r: prefix replaces
instead of intersecting, malformed l: matches all).
